Move masp.rs over to the SDK crate
