+++ apps/src/lib/client/tx.rs
@@ -13,12 +13,12 @@ use namada::core::ledger::governance::cli::offline::{
 use namada::core::ledger::governance::cli::onchain::{
     DefaultProposal, PgfFundingProposal, PgfStewardProposal, ProposalVote,
 };
+use namada::ledger::pos;
 use namada::ledger::wallet::{Wallet, WalletUtils};
-use namada::ledger::{masp, pos};
 use namada::proof_of_stake::parameters::PosParams;
 use namada::proto::Tx;
 use namada::sdk::rpc::{TxBroadcastData, TxResponse};
-use namada::sdk::{signing, tx};
+use namada::sdk::{masp, signing, tx};
 use namada::tendermint_rpc::HttpClient;
 use namada::types::address::{Address, ImplicitAddress};
 use namada::types::dec::Dec;
