+++ apps/src/lib/wallet/cli_utils.rs
@@ -4,8 +4,8 @@ use std::io::{self, Write};
 use borsh::BorshSerialize;
 use itertools::sorted;
 use masp_primitives::zip32::ExtendedFullViewingKey;
-use namada::ledger::masp::find_valid_diversifier;
 use namada::ledger::wallet::{DecryptionError, FindKeyError};
+use namada::sdk::masp::find_valid_diversifier;
 use namada::types::key::{PublicKeyHash, RefTo};
 use namada::types::masp::{MaspValue, PaymentAddress};
 use rand_core::OsRng;
