+++ shared/src/ledger/eth_bridge/bridge_pool.rs
@@ -19,12 +19,12 @@ use super::{block_on_eth_sync, eth_sync_or_exit, BlockOnEthSync};
 use crate::eth_bridge::ethers::abi::AbiDecode;
 use crate::eth_bridge::structs::RelayProof;
 use crate::ledger::args;
-use crate::ledger::masp::{ShieldedContext, ShieldedUtils};
 use crate::ledger::queries::{
     Client, GenBridgePoolProofReq, GenBridgePoolProofRsp, RPC,
 };
 use crate::ledger::wallet::{Wallet, WalletUtils};
 use crate::proto::Tx;
+use crate::sdk::masp::{ShieldedContext, ShieldedUtils};
 use crate::sdk::rpc::{query_wasm_code_hash, validate_amount};
 use crate::sdk::tx::prepare_tx;
 use crate::types::address::Address;
